Structs, headers and tuples initialization/assignment support (#762) (#2017)
Struct, header or tuple can be initialized automatically with a default value of
the suitable type using the syntax `...` or using a mix of explicit values and
default values by using the notation `...` in a list expression
(and struct expression for structs and headers).

Syntax `{ ... }` and `{/* expressionList | kvList */, ... }` can also be used in assignments.

Default values are as proposed in p4-spec pull request #797.

Header stacks and varbits are not supported as a leaf fields of structs
nor elements of tuples if using this kind of initialization/assignment, because in current
version of spec there are no constructs to initialize these types.
If they are used as a leaf fields compiler will report an error.

If header union is used as an element of tuple, compiler will produce warning
that temporary header union variable used for the element initialization may not be
completely initialized, despite setting it invalid.

Frontnend reorganized, Struct initializer now processes declaration initializers
along with assignment statemants, which is the cause of changes in expected
outputs of some tests.